Service layer edit.

+++ src/main/java/com/example/Repositoryes/AuthorizationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorizationRepository  extends JpaRepository<Authorization, Integer> {
+public interface AuthorizationRepository  extends JpaRepository<Authorization, Long> {
 }
