+++ src/main/java/com/example/Repositoryes/StatesArticleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatesArticleRepository extends JpaRepository<StatesArticle, Integer> {
+public interface StatesArticleRepository extends JpaRepository<StatesArticle, Long> {
 }
